feat(scripts): add local GitHub Actions test runner and fix test compilation warnings
- Add run-github-actions-locally.sh script to run Linux/GCC tests locally
- Fix function signatures in loop analysis stubs (added ControlFlowGraph parameter)
- Fix uninitialized variables in ELF relocation tests
- Change test counters from int to size_t to resolve signed/unsigned warnings
- Update all test framework counters for consistency across modules

This ensures clean compilation without warnings and provides a convenient way
to run CI tests locally before pushing changes.

+++ tests/concurrency/bridge_test_common.h
@@ -31,8 +31,8 @@ typedef struct {
     double duration_ms;
 } TestCase;
 
-extern int tests_run;
-extern int tests_passed;
+extern size_t tests_run;
+extern size_t tests_passed;
 
 #define ASSERT(condition, message) \
     do { \
