+++ tests/concurrency/test_bridge_missing_stubs.c
@@ -14,8 +14,8 @@
 // GLOBAL TEST COUNTERS (required by RUN_TEST macro)
 // =============================================================================
 
-int tests_run = 0;
-int tests_passed = 0;
+size_t tests_run = 0;
+size_t tests_passed = 0;
 
 // =============================================================================
 // HELPER FUNCTION IMPLEMENTATIONS
