+++ tests/integration/test_concurrency_tiers_common.h
@@ -28,8 +28,8 @@
 #include "../../runtime/asthra_concurrency_bridge_modular.h"
 
 // Test framework macros and global counters
-extern int tests_run;
-extern int tests_passed;
+extern size_t tests_run;
+extern size_t tests_passed;
 
 #define ASSERT_TRUE(condition) \
     do { \
