+++ tests/performance/test_concurrency_benchmarks_common.h
@@ -25,8 +25,8 @@
 // Global Test Counters
 // ============================================================================
 
-extern int tests_run;
-extern int tests_passed;
+extern size_t tests_run;
+extern size_t tests_passed;
 
 // ============================================================================
 // Test Framework Macros
